requirements.txt
torch
torchvision
flask
pillow
matplotlib